fits: Use latest nom.tam.fits API
Unfortunately, nom.tam.fits (at least until release 1.15.1) re-orders
the header before writing, so that they conform to FITS version 2.
Specifically this puts the `EXTEND` header directly after the `NAXIS*`
headers.

The FITS-PLUS format created in starjava-votable however requires
additional keywords between `NAXIS*` and `EXTEND`. This is a bit
unfortunate, unflexible, and violated the FITS standard when the
FITS-PLUS format was created, but since FITS version 3 the keyword
order does not matter anymore. So, this patch also includes an
alternative `nom.tam.fits.HeaderOrder` class that leaves the order as
it is. This is also the current proposal on the according nom-tam-fits
pull request:

https://github.com/nom-tam-fits/nom-tam-fits/issues/114

Depending on the result of the PR, this patch may be simplified.