ttools: Fix tabletest for use in CI test
This patch adjust tabletest.py so that it may run directly from jython
(if all paths are set accordingly). In this case, the standard method
unittest.main() is called instead of TableTest().runTests(). The
latter is still needed if tabletest.py is run from
JyStiltsTest.java. The choice is made by the testdir global variable
that is preset from JyStiltsTest. If it is not set, we assume the data
file messier.xml in the same directory as the script.

URL: https://github.com/Starlink/starjava/pull/15